DAYD-727 add & refactor 予定の削除ボタンをコンポーネントとして切り出し
