+++ src/components/molecules/TodoCardComponent.tsx
@@ -1,8 +1,5 @@
 import { Plan } from '@/redux/plan/type';
-import { ButtonWithIconComponent } from '@/components/atoms/ButtonWithIconComponent';
-import { MdOutlineDeleteForever } from 'react-icons/md';
-import { FormEvent } from 'react';
-import { useDeletePlanMutation } from '@/redux/plan/slice';
+import { DeletePlanButtonComponent } from './deletePlanButtonComponent';
 
 type Props = {
   todo: Plan;
@@ -10,14 +7,6 @@ type Props = {
 };
 
 export const TodoCardComponent = (props: Props) => {
-  const [deletePlan] = useDeletePlanMutation();
-
-  const handleDeleteSubmit = async (event: FormEvent<HTMLFormElement>, id: number) => {
-    // リロードが走らないように入れている
-    event.preventDefault();
-    deletePlan({ id });
-  };
-
   const className =
     (props.isDragging ? 'w-full' : 'w-[calc(100%_-_2rem)] mx-4') +
     ' ' +
@@ -28,15 +17,7 @@ export const TodoCardComponent = (props: Props) => {
       <div className='text-center text-lg'>{props.todo.title}</div>
       <div className='text-center text-xl flex'>
         <div>{props.todo.processTime}</div>
-        <div>
-          <form id='delete-todo-form' onSubmit={(event) => handleDeleteSubmit(event, props.todo.id)}>
-            <ButtonWithIconComponent
-              type='submit'
-              icon={<MdOutlineDeleteForever />}
-              size={1.5}
-            ></ButtonWithIconComponent>
-          </form>
-        </div>
+        <DeletePlanButtonComponent size={1.5} planId={props.todo.id} />
       </div>
     </div>
   );
